document-intelligence-lab-main/requirements.txt
streamlit>=1.35
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0
Pillow>=10.0
requests>=2.31
pymupdf>=1.23
numpy>=1.24
# Docling is the core parser. Install may take time because it pulls document models/dependencies.
docling>=2.0
